1번 회원을 관리자로 지정, 게시물 수정/삭제 권한 체크 exit d
fdsd

fdsc
c
sa
sad

+++ src/main/java/com/sbs/untact/controller/UsrHomeController.java
@@ -76,7 +76,13 @@ public class UsrHomeController {
 
 	@RequestMapping("/usr/article/doDelete")
 	@ResponseBody
-	public ResultData doDelete(Integer id) {
+	public ResultData doDelete(Integer id, HttpSession session) {
+		int loginedMemberId = Util.getAsInt(session.getAttribute("loginedMemberId"), 0);
+		
+		if (loginedMemberId == 0) {
+			return new ResultData("F-2", "로그인부터 해주세요.");
+		}
+		
 		if (id == null) {
 			return new ResultData("F-1", "아이디를 입력해주세요.");
 		}
@@ -84,13 +90,25 @@ public class UsrHomeController {
 		if (article == null) {
 			return new ResultData("F-1", "해당 게시물은 존재하지 않습니다.");
 		}
+		
+		ResultData actorCanDeleteRd = articleService.getActorCanDeleteRd(article, loginedMemberId);
+		
+		if (actorCanDeleteRd.isFail()) {
+			return actorCanDeleteRd;
+		}
 		return articleService.deleteArticle(id);
 	}
 
 
 	@RequestMapping("/usr/article/doModify")
 	@ResponseBody
-	public ResultData doModify(Integer id, String title, String body) {
+	public ResultData doModify(Integer id, String title, String body, HttpSession session) {
+		int loginedMemberId = Util.getAsInt(session.getAttribute("loginedMemberId"), 0);
+		
+		if (loginedMemberId == 0) {
+			return new ResultData("F-2", "로그인부터 해주세요.");
+		}
+		
 		if (id == null) {
 			return new ResultData("F-1", "아이디를 입력해주세요.");
 		}
@@ -102,10 +120,16 @@ public class UsrHomeController {
 		}
 		
 		Article article = articleService.getArticle(id);
+		
 		if (article == null) {
 			return new ResultData("F-1", "해당 게시물은 존재하지 않습니다.");
 		}
 		
+		ResultData actorCanModifyRd = articleService.getActorCanModifyRd(article, loginedMemberId);
+		
+		if (actorCanModifyRd.isFail()) {
+			return actorCanModifyRd;
+		}
 		return articleService.modifyArticle(id, title, body);
 	}
 }
